+++ tests/test_ensembles.py
@@ -23,6 +23,7 @@ import pytest
 import xarray as xr
 
 from xclim import ensembles
+from xclim.indices.stats import get_dist
 from xclim.testing import open_dataset
 
 
@@ -467,3 +468,126 @@ class TestEnsembleReduction:
             )
             assert sel1 == sel2
             assert sel1 == sel3
+
+
+# ## Tests for Robustness ##
+@pytest.fixture(params=[True, False])
+def robust_data(request):
+    norm = get_dist("norm")
+    ref = np.tile(
+        np.array(
+            [
+                norm.rvs(loc=274, scale=0.8, size=(40,), random_state=r)
+                for r in [101083, 19377, 473820, 483625]
+            ]
+        ),
+        (4, 1, 1),
+    )
+    fut = np.array(
+        [
+            [
+                norm.rvs(loc=loc, scale=sc, size=(40,), random_state=r)
+                for loc, sc, r in shps
+            ]
+            for shps in (
+                [
+                    (274.0, 0.7, 176378),
+                    (274.0, 0.6, 839789),
+                    (274.0, 0.7, 393239),
+                    (275.6, 1.1, 747390),
+                ],  # 3 no change, 1 positive change
+                [
+                    (272.5, 1.2, 743920),
+                    (272.4, 0.8, 138489),
+                    (275.5, 0.8, 673683),
+                    (275.6, 1.1, 969383),
+                ],  # 2 neg change
+                [
+                    (275.6, 0.8, 696857),
+                    (275.8, 1.2, 379949),
+                    (276.5, 0.8, 268395),
+                    (277.6, 1.1, 456544),
+                ],  # All pos change
+                [
+                    (np.nan, 0.3, 746323),
+                    (np.nan, 1.2, 5643723),
+                    (275.5, 0.8, 118294),
+                    (275.6, 1.1, 574732),
+                ],  # Some NaN
+            )
+        ]
+    )
+    ref = xr.DataArray(ref, dims=("lon", "realization", "time"), name="tas")
+    ref["time"] = xr.cftime_range("2000-01-01", periods=40, freq="YS")
+    fut = xr.DataArray(fut, dims=("lon", "realization", "time"), name="tas")
+    fut["time"] = xr.cftime_range("2040-01-01", periods=40, freq="YS")
+    if request.param:
+        ref = ref.chunk({"lon": 1}).to_dataset()
+        fut = fut.chunk({"lon": 1}).to_dataset()
+    return ref, fut
+
+
+@pytest.mark.parametrize(
+    "test,exp_chng,exp_sign,kws",
+    [
+        ("ttest", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {}),
+        ("welch-ttest", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {}),
+        ("threshold", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {"rel_thresh": 0.002}),
+        (
+            "threshold",
+            [0, 0, 0.5, 0],
+            [np.nan, np.nan, 1, np.nan],
+            {"abs_thresh": 2},
+        ),
+        (None, [1, 1, 1, 1], [0.5, 0.5, 1, 1], {}),
+    ],
+)
+def test_change_significance(robust_data, test, exp_chng, exp_sign, kws):
+    ref, fut = robust_data
+    chng, sign = ensembles.change_significance(fut, ref, test=test, **kws)
+    assert chng.attrs["test"] == str(test)
+    if isinstance(ref, xr.Dataset):
+        chng = chng.tas
+        sign = sign.tas
+    np.testing.assert_array_almost_equal(chng, exp_chng)
+    np.testing.assert_array_almost_equal(sign, exp_sign)
+
+
+def test_change_significance_delta(robust_data):
+    ref, fut = robust_data
+    delta = fut.mean("time") - ref.mean("time")
+    chng, sign = ensembles.change_significance(delta, test="threshold", abs_thresh=2)
+    if isinstance(ref, xr.Dataset):
+        chng = chng.tas
+        sign = sign.tas
+    np.testing.assert_array_equal(chng, [0, 0, 0.5, 0])
+    np.testing.assert_array_equal(sign, [np.nan, np.nan, 1, np.nan])
+
+
+def test_robustness_coefficient():
+    # High
+    ref = xr.DataArray([274, 275, 274.5, 276, 274.3, 273.3], dims=("time",), name="tas")
+    fut = xr.DataArray(
+        [
+            [277, 277.1, 278, 278.4, 278.1, 276.9],
+            [275, 275.8, 276, 275.2, 276.2, 275.7],
+        ],
+        dims=("realization", "time"),
+        name="tas",
+    )
+    R = ensembles.robustness_coefficient(fut, ref)
+    np.testing.assert_almost_equal(R, 0.91972477)
+
+    fut = xr.DataArray(
+        [
+            [277, 277.1, 278, 278.4, 278.1, 276.9],
+            [274, 274.8, 273.7, 274.2, 273.9, 274.5],
+        ],
+        dims=("realization", "time"),
+        name="tas",
+    )
+    R = ensembles.robustness_coefficient(fut, ref)
+    np.testing.assert_almost_equal(R, 0.83743842)
+
+    R = ensembles.robustness_coefficient(fut.to_dataset(), ref.to_dataset())
+    np.testing.assert_almost_equal(R.tas, 0.83743842)