+++ tests/test_temperature.py
@@ -6,6 +6,7 @@ import xarray as xr
 
 from xclim import atmos
 from xclim.core.calendar import percentile_doy
+from xclim.core.options import set_options
 from xclim.testing import open_dataset
 
 K2C = 273.15
@@ -1186,3 +1187,30 @@ def test_freshet_start(tas_series):
         tas_series(np.arange(-50, 350) + 274, start="1/1/2000"), freq="YS"
     )
     assert out[0] == 51
+
+
+def test_degree_days_depassment_date():
+    tas = open_dataset("FWI/GFWED_sample_2017.nc").tas
+    tas.attrs.update(
+        cell_methods="time: mean within days", standard_name="air_temperature"
+    )
+
+    out = atmos.degree_days_depassment_date(
+        tas=tas,
+        thresh="4 degC",
+        op=">",
+        sum_thresh="200 K days",
+    )
+    np.testing.assert_array_equal(out, np.array([[153, 136, 9, 6]]).T)
+    assert "tmean > 4 degc" in out.attrs["description"]
+
+    with set_options(check_missing="skip"):
+        out = atmos.degree_days_depassment_date(
+            tas=tas,
+            thresh="4 degC",
+            op=">",
+            sum_thresh="1500 K days",
+            start_date="07-02",
+            freq="YS",
+        )
+        np.testing.assert_array_equal(out, np.array([[np.nan, 280, 241, 244]]).T)